Started UI for MQTT broker selection; some reorganization.
